% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.R
\name{load_collect_wgs_metrics}
\alias{load_collect_wgs_metrics}
\title{Load CollectWgsMetrics Output}
\usage{
load_collect_wgs_metrics(infile)
}
\arguments{
\item{infile}{Path to CollectWgsMetrics output.}
}
\description{
Loads the output file of CollectWgsMetrics. This will load the
METRICS CLASS portion of the output.
}

